requirements.txt
playwright==1.44.0
customtkinter>=5.2.0
openpyxl>=3.1.0
pystray>=0.19.0
pillow>=10.0.0
